requirements.dev.txt
pytest-django
pytest-sugar
pytest-cache
coverage
pycodestyle
